requirements.txt
pyyaml
requests
py-getch
rich